:recycle: Create a ProductMapper
