+++ src/main/java/org/example/reactive_general_market/src/product/application/dto/CreatedProductDto.java
@@ -1,4 +1,4 @@
-package org.example.reactive_general_market.src.product.infrastructure.dto;
+package org.example.reactive_general_market.src.product.application.dto;
 
 import java.util.UUID;
