--- src/main/java/org/example/reactive_general_market/src/product/infrastructure/dto/ErrorResponse.java
@@ -1,5 +1,0 @@
-package org.example.reactive_general_market.src.product.infrastructure.dto;
-
-public record ErrorResponse(
-    String message
-) {}
