+++ src/main/java/org/example/reactive_general_market/src/product/infrastructure/handler/ProductHandler.java
@@ -2,12 +2,9 @@ package org.example.reactive_general_market.src.product.infrastructure.handler;
 
 import org.example.reactive_general_market.src.product.application.CreateProduct;
 import org.example.reactive_general_market.src.product.application.FindPaginatedProducts;
-import org.example.reactive_general_market.src.product.application.model.ProductPage;
-import org.example.reactive_general_market.src.product.application.model.ProductsResultDto;
-import org.example.reactive_general_market.src.product.domain.model.Product;
-import org.example.reactive_general_market.src.product.infrastructure.dto.CreatedProductDto;
-import org.example.reactive_general_market.src.product.infrastructure.dto.ErrorResponse;
-import org.springframework.data.domain.PageImpl;
+import org.example.reactive_general_market.src.product.application.dto.CreatedProductDto;
+import org.example.reactive_general_market.src.product.infrastructure.model.ErrorResponse;
+import org.example.reactive_general_market.src.product.infrastructure.mapper.ProductMapper;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.ServerRequest;
@@ -37,22 +34,9 @@ public class ProductHandler {
     final Pageable pageable = Pageable.ofSize(size).withPage(page);
 
     return findPaginatedProducts.execute(pageable)
-        .flatMap(productsResultDto -> ServerResponse.ok().bodyValue(toPage(productsResultDto, pageable)))
+        .flatMap(productsResultDto ->
+            ServerResponse.ok().bodyValue(ProductMapper.toPage(productsResultDto, pageable))
+        )
         .switchIfEmpty(ServerResponse.notFound().build());
-  }
-
-  private ProductPage toPage(ProductsResultDto productsResultDto, Pageable pageable) {
-    final PageImpl<Product> productPageImpl = new PageImpl<>(productsResultDto.products(), pageable,
-        productsResultDto.totalProducts());
-    return new ProductPage(
-        productsResultDto.products(),
-        productPageImpl.getNumberOfElements(),
-        productPageImpl.getTotalPages(),
-        productPageImpl.getNumber(),
-        productPageImpl.getSize(),
-        productPageImpl.isFirst(),
-        productPageImpl.isLast(),
-        productPageImpl.isEmpty()
-    );
   }
 }
