+++ src/main/java/org/example/reactive_general_market/src/product/infrastructure/model/ErrorResponse.java
@@ -1,0 +1,5 @@
+package org.example.reactive_general_market.src.product.infrastructure.model;
+
+public record ErrorResponse(
+    String message
+) {}
